split off move fsm
