+++ src/software/ai/hl/stp/tactic/move_tactic_test.cpp
@@ -59,40 +59,3 @@ TEST(MoveTacticTest, test_calculate_robot_cost)
 
     EXPECT_EQ(5 / world.field().totalXLength(), tactic.calculateRobotCost(robot, world));
 }
-
-TEST(MoveFSMTest, test_transitions)
-{
-    World world = ::TestUtil::createBlankTestingWorld();
-    Robot robot = ::TestUtil::createRobotAtPos(Point(-2, -3));
-    MoveFSM::ControlParams control_params{.destination       = Point(2, 3),
-                                          .final_orientation = Angle::half(),
-                                          .final_speed       = 0.0};
-
-    boost::sml::sm<MoveFSM> fsm;
-    EXPECT_TRUE(fsm.is(boost::sml::state<MoveFSM::idle_state>));
-    fsm.process_event(MoveFSM::Update{
-        .control_params = control_params,
-        .common         = TacticUpdate{.robot      = robot,
-                               .world      = world,
-                               .set_intent = [](std::unique_ptr<Intent>) {}}});
-    // robot far from destination
-    EXPECT_TRUE(fsm.is(boost::sml::state<MoveFSM::move_state>));
-    robot = ::TestUtil::createRobotAtPos(Point(2, 2));
-    fsm.process_event(MoveFSM::Update{
-        .control_params = control_params,
-        .common         = TacticUpdate{.robot      = robot,
-                               .world      = world,
-                               .set_intent = [](std::unique_ptr<Intent>) {}}});
-    // robot close to destination
-    EXPECT_TRUE(fsm.is(boost::sml::state<MoveFSM::move_state>));
-    robot.updateState(
-        RobotState(Point(2, 3), Vector(), Angle::half(), AngularVelocity::zero()),
-        Timestamp::fromSeconds(0));
-    fsm.process_event(MoveFSM::Update{
-        .control_params = control_params,
-        .common         = TacticUpdate{.robot      = robot,
-                               .world      = world,
-                               .set_intent = [](std::unique_ptr<Intent>) {}}});
-    // robot at destination and facing the right way
-    EXPECT_TRUE(fsm.is(boost::sml::X));
-}
